Remove Resource model
