+++ model/instances_test.go
@@ -1,12 +1,16 @@
 package model
 
 import (
+	"fmt"
+	"reflect"
+	"strings"
 	"testing"
 
 	"golang.org/x/net/context"
 
 	"github.com/axsh/openvdc/internal/unittest"
 	"github.com/axsh/openvdc/model/backend"
+	"github.com/golang/protobuf/ptypes/timestamp"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -41,9 +45,6 @@ func TestCreateInstance(t *testing.T) {
 
 	var err error
 	_, err = Instances(context.Background()).Create(n)
-	assert.Equal(ErrInstanceMissingResource, err)
-	n.ResourceId = "r-xxxx"
-	_, err = Instances(context.Background()).Create(n)
 	assert.Equal(ErrBackendNotInContext, err)
 
 	withConnect(t, func(ctx context.Context) {
@@ -56,9 +57,13 @@ func TestCreateInstance(t *testing.T) {
 func TestFindInstance(t *testing.T) {
 	assert := assert.New(t)
 	n := &Instance{
-		SlaveId:    "xxx",
-		ResourceId: "r-xxxx",
-	}
+		SlaveId: "xxx",
+		Template: &Template{
+			Item: &Template_Lxc{
+				Lxc: lxcTmpl1,
+			},
+		}}
+
 	_, err := Instances(context.Background()).FindByID("i-xxxxx")
 	assert.Equal(ErrBackendNotInContext, err)
 
@@ -81,9 +86,12 @@ func TestUpdateStateInstance(t *testing.T) {
 
 	withConnect(t, func(ctx context.Context) {
 		n := &Instance{
-			SlaveId:    "xxx",
-			ResourceId: "r-xxxx",
-		}
+			SlaveId: "xxx",
+			Template: &Template{
+				Item: &Template_Lxc{
+					Lxc: lxcTmpl1,
+				},
+			}}
 		got, err := Instances(ctx).Create(n)
 		assert.NoError(err)
 		assert.Equal(InstanceState_REGISTERED, got.GetLastState().State)
@@ -109,20 +117,23 @@ func TestUpdateInstance(t *testing.T) {
 
 	withConnect(t, func(ctx context.Context) {
 		n := &Instance{
-			SlaveId:    "xxx",
-			ResourceId: "r-xxxx",
-		}
+			SlaveId: "xxx",
+			Template: &Template{
+				Item: &Template_Lxc{
+					Lxc: lxcTmpl1,
+				},
+			}}
 		err = Instances(ctx).Update(n)
 		assert.Error(err)
 		assert.Equal(ErrInvalidID, err, "Empty ID object should get ErrInvalidID")
 
 		got, err := Instances(ctx).Create(n)
 		assert.NoError(err)
-		got.ResourceId = "r-yyyy"
+		got.Template.GetLxc().Vcpu = 100
 		err = Instances(ctx).Update(got)
 		got2, err := Instances(ctx).FindByID(got.Id)
 		assert.NoError(err)
-		assert.Equal(got.ResourceId, got2.ResourceId)
+		assert.Equal(int32(100), got2.Template.GetLxc().Vcpu)
 	})
 }
 
@@ -171,4 +182,64 @@ func TestInstanceState_ValidateGoalState(t *testing.T) {
 	s.State = InstanceState_TERMINATED
 	assert.Error(s.ValidateGoalState(InstanceState_TERMINATED))
 	assert.Error(s.ValidateGoalState(InstanceState_RUNNING))
+}
+
+var lxcTmpl1 = &LxcTemplate{
+	Vcpu:     1,
+	MemoryGb: 10,
+	LxcImage: &LxcTemplate_Image{
+		DownloadUrl: "http://example.com/image.raw",
+		Chksum:      "1234567890abcdef",
+	},
+}
+
+var res1 = &Instance{
+	CreatedAt: new(timestamp.Timestamp),
+	Template: &Template{
+		TemplateUri: "https://example.com/xxx.json",
+		Item: &Template_Lxc{
+			Lxc: &LxcTemplate{
+				Vcpu:     1,
+				MemoryGb: 10,
+				LxcImage: &LxcTemplate_Image{
+					DownloadUrl: "http://example.com/image.raw",
+					Chksum:      "1234567890abcdef",
+				},
+			}}}}
+
+func TestInstance_ResourceTemplate(t *testing.T) {
+	assert := assert.New(t)
+	assert.Equal("vm/lxc", res1.ResourceTemplate().ResourceName())
+}
+
+func ExampleResourceTemplate_reflection() {
+	// Normal type assertion
+	rl, ok := res1.Template.Item.(*Template_Lxc)
+	fmt.Println(rl.Lxc, ok)
+	// Using reflection API
+	v := reflect.ValueOf(res1.Template.Item)
+	fmt.Println(v.Kind())
+	fmt.Println(v.Type().String())
+	fmt.Println(
+		"ConvertibleTo(*Template_Lxc) ->",
+		v.Type().ConvertibleTo(reflect.TypeOf((*Template_Lxc)(nil))),
+	)
+	fmt.Println(
+		"ConvertibleTo(*Template_Null) ->",
+		v.Type().ConvertibleTo(reflect.TypeOf((*Template_Null)(nil))),
+	)
+	fieldName := strings.TrimPrefix(v.Type().String(), "*model.Template_")
+	fmt.Println(fieldName)
+	f := v.Elem().FieldByName(fieldName)
+	fmt.Println(f.Type().String())
+	fmt.Println(f.MethodByName("ResourceName").Call([]reflect.Value{})[0])
+	// Output:
+	// vcpu:1 memory_gb:10 lxc_image:<download_url:"http://example.com/image.raw" chksum:"1234567890abcdef" >  true
+	// ptr
+	// *model.Template_Lxc
+	// ConvertibleTo(*Template_Lxc) -> true
+	// ConvertibleTo(*Template_Null) -> false
+	// Lxc
+	// *model.LxcTemplate
+	// vm/lxc
 }
