+++ scheduler/mesos.go
@@ -116,22 +116,15 @@ func (sched *VDCScheduler) processOffers(driver sched.SchedulerDriver, offers []
 	}
 
 	findMatching := func(i *model.Instance) *mesos.Offer {
+		log := log.WithField("instance_id", i.GetId())
 		for _, offer := range offers {
 			hypervisorName := getHypervisorName(offer)
 			if hypervisorName == "" {
 				continue
 			}
 
-			r, err := i.Resource(ctx)
-			if err != nil {
-				log.WithError(err).WithFields(log.Fields{
-					"instance_id": i.GetId(),
-					"resource_id": i.GetResourceId(),
-				}).Error("Failed to retrieve resource object")
-				continue
-			}
 			// TODO: Avoid type switch to find template types.
-			switch t := r.GetTemplate().GetItem(); t.(type) {
+			switch t := i.GetTemplate().GetItem(); t.(type) {
 			case *model.Template_Lxc:
 				if hypervisorName == "lxc" {
 					return offer
@@ -176,16 +169,6 @@ func (sched *VDCScheduler) processOffers(driver sched.SchedulerDriver, offers []
 			},
 		}
 
-		r, err := i.Resource(ctx)
-
-		if err != nil {
-			log.WithError(err).WithFields(log.Fields{
-				"instance_id": i.GetId(),
-				"resource_id": i.GetResourceId(),
-			}).Error("Failed to retrieve resource object")
-			continue
-		}
-
 		taskId := util.NewTaskID(i.GetId())
 		task := &mesos.TaskInfo{
 			Name:     proto.String("VDC" + "_" + taskId.GetValue()),
@@ -194,8 +177,8 @@ func (sched *VDCScheduler) processOffers(driver sched.SchedulerDriver, offers []
 			Data:     []byte("instance_id=" + i.GetId()),
 			Executor: executor,
 			Resources: []*mesos.Resource{
-				util.NewScalarResource("cpus", float64(r.GetTemplate().GetLxc().GetVcpu())),
-				util.NewScalarResource("mem", float64(r.GetTemplate().GetLxc().GetMemoryGb()*1024)),
+				util.NewScalarResource("cpus", float64(i.GetTemplate().GetLxc().GetVcpu())),
+				util.NewScalarResource("mem", float64(i.GetTemplate().GetLxc().GetMemoryGb()*1024)),
 			},
 		}
